Powerups view half done - Blessings now show the correct amount when miracles are active - Condensed all modal handlers into one class  - Only one modal can be open at once, no need to make each modal type distinct - Added powerups view  - Gives top-level information about powerups  - Non-functional powerup select menu - Added missing rarities to clone stats page

+++ src/main/java/dev/boarbot/entities/boaruser/ProfileData.java
@@ -11,7 +11,7 @@ public record ProfileData(
     int uniqueBoars,
     int numSkyblock,
     int streak,
-    int blessings,
+    long blessings,
     int streakBless,
     int uniqueBless,
     int questBless,
